+++ app/src/main/kotlin/com/simplemobiletools/calendar/pro/fragments/WeekFragment.kt
@@ -1,15 +1,12 @@
 package com.simplemobiletools.calendar.pro.fragments
 
+import android.annotation.SuppressLint
 import android.content.Intent
 import android.content.res.Resources
-import android.graphics.Rect
 import android.graphics.drawable.ColorDrawable
 import android.os.Bundle
 import android.util.Range
-import android.view.LayoutInflater
-import android.view.MotionEvent
-import android.view.View
-import android.view.ViewGroup
+import android.view.*
 import android.widget.ImageView
 import android.widget.RelativeLayout
 import android.widget.TextView
@@ -17,10 +14,7 @@ import androidx.collection.LongSparseArray
 import androidx.fragment.app.Fragment
 import com.simplemobiletools.calendar.pro.R
 import com.simplemobiletools.calendar.pro.activities.EventActivity
-import com.simplemobiletools.calendar.pro.extensions.config
-import com.simplemobiletools.calendar.pro.extensions.eventsHelper
-import com.simplemobiletools.calendar.pro.extensions.seconds
-import com.simplemobiletools.calendar.pro.extensions.touch
+import com.simplemobiletools.calendar.pro.extensions.*
 import com.simplemobiletools.calendar.pro.helpers.*
 import com.simplemobiletools.calendar.pro.helpers.Formatter
 import com.simplemobiletools.calendar.pro.interfaces.WeekFragmentListener
@@ -38,64 +32,92 @@ import org.joda.time.Days
 import java.util.*
 
 class WeekFragment : Fragment(), WeeklyCalendar {
-    private val CLICK_DURATION_THRESHOLD = 150
     private val PLUS_FADEOUT_DELAY = 5000L
+    private val MIN_SCALE_FACTOR = 0.3f
+    private val MAX_SCALE_FACTOR = 5f
+    private val MIN_SCALE_DIFFERENCE = 0.02f
+    private val SCALE_RANGE = MAX_SCALE_FACTOR - MIN_SCALE_FACTOR
 
-    var mListener: WeekFragmentListener? = null
-    private var mWeekTimestamp = 0L
-    private var mRowHeight = 0f
-    private var minScrollY = -1
-    private var maxScrollY = -1
+    var listener: WeekFragmentListener? = null
+    private var weekTimestamp = 0L
+    private var rowHeight = 0f
     private var todayColumnIndex = -1
-    private var clickStartTime = 0L
     private var primaryColor = 0
     private var lastHash = 0
+    private var prevScaleSpanY = 0f
+    private var scaleCenterPercent = 0f
+    private var defaultRowHeight = 0f
+    private var screenHeight = 0
+    private var rowHeightsAtScale = 0f
+    private var prevScaleFactor = 0f
     private var mWasDestroyed = false
     private var isFragmentVisible = false
     private var wasFragmentInit = false
     private var wasExtraHeightAdded = false
     private var dimPastEvents = true
+    private var wasScaled = false
     private var selectedGrid: View? = null
     private var currentTimeView: ImageView? = null
-    private var events = ArrayList<Event>()
     private var allDayHolders = ArrayList<RelativeLayout>()
     private var allDayRows = ArrayList<HashSet<Int>>()
+    private var currEvents = ArrayList<Event>()
     private var eventTypeColors = LongSparseArray<Int>()
     private var eventTimeRanges = LinkedHashMap<String, ArrayList<EventWeeklyView>>()
 
     private lateinit var inflater: LayoutInflater
     private lateinit var mView: View
-    private lateinit var mScrollView: MyScrollView
-    private lateinit var mCalendar: WeeklyCalendarImpl
-    private lateinit var mRes: Resources
-    private lateinit var mConfig: Config
+    private lateinit var scrollView: MyScrollView
+    private lateinit var res: Resources
+    private lateinit var config: Config
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        mRes = context!!.resources
-        mConfig = context!!.config
-        mRowHeight = mRes.getDimension(R.dimen.weekly_view_row_height)
-        minScrollY = (mRowHeight * mConfig.startWeeklyAt).toInt()
-        mWeekTimestamp = arguments!!.getLong(WEEK_START_TIMESTAMP)
-        dimPastEvents = mConfig.dimPastEvents
+        res = context!!.resources
+        config = context!!.config
+        rowHeight = context!!.getWeeklyViewItemHeight()
+        defaultRowHeight = res.getDimension(R.dimen.weekly_view_row_height)
+        weekTimestamp = arguments!!.getLong(WEEK_START_TIMESTAMP)
+        dimPastEvents = config.dimPastEvents
         primaryColor = context!!.getAdjustedPrimaryColor()
         allDayRows.add(HashSet())
-        mCalendar = WeeklyCalendarImpl(this, context!!)
     }
 
+    @SuppressLint("ClickableViewAccessibility")
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
         this.inflater = inflater
 
-        mView = inflater.inflate(R.layout.fragment_week, container, false)
-        mScrollView = mView.week_events_scrollview
-        mScrollView.setOnScrollviewListener(object : MyScrollView.ScrollViewListener {
+        val fullHeight = context!!.getWeeklyViewItemHeight().toInt() * 24
+        mView = inflater.inflate(R.layout.fragment_week, container, false).apply {
+            scrollView = week_events_scrollview
+            week_horizontal_grid_holder.layoutParams.height = fullHeight
+            week_events_columns_holder.layoutParams.height = fullHeight
+
+            val scaleDetector = getViewScaleDetector()
+            scrollView.setOnTouchListener { view, motionEvent ->
+                scaleDetector.onTouchEvent(motionEvent)
+                if (motionEvent.action == MotionEvent.ACTION_UP && wasScaled) {
+                    scrollView.isScrollable = true
+                    wasScaled = false
+                    true
+                } else {
+                    false
+                }
+            }
+        }
+
+        scrollView.setOnScrollviewListener(object : MyScrollView.ScrollViewListener {
             override fun onScrollChanged(scrollView: MyScrollView, x: Int, y: Int, oldx: Int, oldy: Int) {
                 checkScrollLimits(y)
             }
         })
 
-        mScrollView.onGlobalLayout {
-            updateScrollY(Math.max(mListener?.getCurrScrollY() ?: 0, minScrollY))
+        scrollView.onGlobalLayout {
+            if (fullHeight < scrollView.height) {
+                scrollView.layoutParams.height = fullHeight - context!!.resources.getDimension(R.dimen.one_dp).toInt()
+            }
+
+            val initialScrollY = (rowHeight * config.startWeeklyAt).toInt()
+            updateScrollY(Math.max(listener?.getCurrScrollY() ?: 0, initialScrollY))
         }
 
         wasFragmentInit = true
@@ -105,28 +127,13 @@ class WeekFragment : Fragment(), WeeklyCalendar {
     override fun onResume() {
         super.onResume()
         context!!.eventsHelper.getEventTypes(activity!!, false) {
-            it.map { eventTypeColors.put(it.id!!, it.color) }
+            it.map {
+                eventTypeColors.put(it.id!!, it.color)
+            }
         }
 
         setupDayLabels()
         updateCalendar()
-
-        mScrollView.onGlobalLayout {
-            if (context == null) {
-                return@onGlobalLayout
-            }
-
-            minScrollY = (mRowHeight * mConfig.startWeeklyAt).toInt()
-            maxScrollY = (mRowHeight * mConfig.endWeeklyAt).toInt()
-
-            val bounds = Rect()
-            week_events_holder.getGlobalVisibleRect(bounds)
-            maxScrollY -= bounds.bottom - bounds.top
-            if (minScrollY > maxScrollY)
-                maxScrollY = -1
-
-            checkScrollLimits(mScrollView.scrollY)
-        }
     }
 
     override fun onPause() {
@@ -143,25 +150,33 @@ class WeekFragment : Fragment(), WeeklyCalendar {
         super.setMenuVisibility(menuVisible)
         isFragmentVisible = menuVisible
         if (isFragmentVisible && wasFragmentInit) {
-            mListener?.updateHoursTopMargin(mView.week_top_holder.height)
-            checkScrollLimits(mScrollView.scrollY)
+            listener?.updateHoursTopMargin(mView.week_top_holder.height)
+            checkScrollLimits(scrollView.scrollY)
+
+            // fix some glitches like at swiping from a fully scaled out fragment will all-day events to an empty one
+            val fullFragmentHeight = (listener?.getFullFragmentHeight() ?: 0) - mView.week_top_holder.height
+            if (scrollView.height < fullFragmentHeight) {
+                config.weeklyViewItemHeightMultiplier = fullFragmentHeight / 24 / defaultRowHeight
+                updateViewScale()
+                listener?.updateRowHeight(rowHeight.toInt())
+            }
         }
     }
 
     fun updateCalendar() {
-        mCalendar.updateWeeklyCalendar(mWeekTimestamp)
+        WeeklyCalendarImpl(this, context!!).updateWeeklyCalendar(weekTimestamp)
     }
 
     private fun setupDayLabels() {
-        var curDay = Formatter.getDateTimeFromTS(mWeekTimestamp)
-        val textColor = mConfig.textColor
+        var curDay = Formatter.getDateTimeFromTS(weekTimestamp)
+        val textColor = config.textColor
         val todayCode = Formatter.getDayCodeFromDateTime(DateTime())
         for (i in 0..6) {
             val dayCode = Formatter.getDayCodeFromDateTime(curDay)
-            val dayLetters = mRes.getStringArray(R.array.week_day_letters).toMutableList() as ArrayList<String>
+            val dayLetters = res.getStringArray(R.array.week_day_letters).toMutableList() as ArrayList<String>
             val dayLetter = dayLetters[curDay.dayOfWeek - 1]
 
-            mView.findViewById<TextView>(mRes.getIdentifier("week_day_label_$i", "id", context!!.packageName)).apply {
+            mView.findViewById<TextView>(res.getIdentifier("week_day_label_$i", "id", context!!.packageName)).apply {
                 text = "$dayLetter\n${curDay.dayOfMonth}"
                 setTextColor(if (todayCode == dayCode) primaryColor else textColor)
                 if (todayCode == dayCode) {
@@ -173,12 +188,8 @@ class WeekFragment : Fragment(), WeeklyCalendar {
     }
 
     private fun checkScrollLimits(y: Int) {
-        if (minScrollY != -1 && y < minScrollY) {
-            mScrollView.scrollY = minScrollY
-        } else if (maxScrollY != -1 && y > maxScrollY) {
-            mScrollView.scrollY = maxScrollY
-        } else if (isFragmentVisible) {
-            mListener?.scrollTo(y)
+        if (isFragmentVisible) {
+            listener?.scrollTo(y)
         }
     }
 
@@ -186,45 +197,89 @@ class WeekFragment : Fragment(), WeeklyCalendar {
         (0..6).map { getColumnWithId(it) }
                 .forEachIndexed { index, layout ->
                     layout.removeAllViews()
+                    val gestureDetector = getViewGestureDetector(layout, index)
+
                     layout.setOnTouchListener { view, motionEvent ->
-                        checkGridClick(motionEvent, index, layout)
+                        gestureDetector.onTouchEvent(motionEvent)
                         true
                     }
                 }
     }
 
-    private fun checkGridClick(event: MotionEvent, index: Int, view: ViewGroup) {
-        when (event.action) {
-            MotionEvent.ACTION_DOWN -> clickStartTime = System.currentTimeMillis()
-            MotionEvent.ACTION_UP -> {
-                if (System.currentTimeMillis() - clickStartTime < CLICK_DURATION_THRESHOLD) {
-                    selectedGrid?.animation?.cancel()
-                    selectedGrid?.beGone()
+    private fun getViewGestureDetector(view: ViewGroup, index: Int): GestureDetector {
+        return GestureDetector(context, object : GestureDetector.SimpleOnGestureListener() {
+            override fun onSingleTapUp(event: MotionEvent): Boolean {
+                selectedGrid?.animation?.cancel()
+                selectedGrid?.beGone()
 
-                    val hour = (event.y / mRowHeight).toInt()
-                    selectedGrid = (inflater.inflate(R.layout.week_grid_item, null, false) as ImageView).apply {
-                        view.addView(this)
-                        background = ColorDrawable(primaryColor)
-                        layoutParams.width = view.width
-                        layoutParams.height = mRowHeight.toInt()
-                        y = hour * mRowHeight
-                        applyColorFilter(primaryColor.getContrastColor())
+                val hour = (event.y / rowHeight).toInt()
+                selectedGrid = (inflater.inflate(R.layout.week_grid_item, null, false) as ImageView).apply {
+                    view.addView(this)
+                    background = ColorDrawable(primaryColor)
+                    layoutParams.width = view.width
+                    layoutParams.height = rowHeight.toInt()
+                    y = hour * rowHeight
+                    applyColorFilter(primaryColor.getContrastColor())
 
-                        setOnClickListener {
-                            val timestamp = mWeekTimestamp + index * DAY_SECONDS + hour * 60 * 60
-                            Intent(context, EventActivity::class.java).apply {
-                                putExtra(NEW_EVENT_START_TS, timestamp)
-                                putExtra(NEW_EVENT_SET_HOUR_DURATION, true)
-                                startActivity(this)
-                            }
-                        }
-                        animate().alpha(0f).setStartDelay(PLUS_FADEOUT_DELAY).withEndAction {
-                            beGone()
+                    setOnClickListener {
+                        val timestamp = weekTimestamp + index * DAY_SECONDS + hour * 60 * 60
+                        Intent(context, EventActivity::class.java).apply {
+                            putExtra(NEW_EVENT_START_TS, timestamp)
+                            putExtra(NEW_EVENT_SET_HOUR_DURATION, true)
+                            startActivity(this)
                         }
                     }
+
+                    animate().setStartDelay(PLUS_FADEOUT_DELAY).alpha(0f).withEndAction {
+                        beGone()
+                    }
                 }
+                return super.onSingleTapUp(event)
             }
-        }
+        })
+    }
+
+    private fun getViewScaleDetector(): ScaleGestureDetector {
+        return ScaleGestureDetector(context, object : ScaleGestureDetector.SimpleOnScaleGestureListener() {
+            override fun onScale(detector: ScaleGestureDetector): Boolean {
+                val percent = (prevScaleSpanY - detector.currentSpanY) / screenHeight
+                prevScaleSpanY = detector.currentSpanY
+
+                val wantedFactor = config.weeklyViewItemHeightMultiplier - (SCALE_RANGE * percent)
+                var newFactor = Math.max(Math.min(wantedFactor, MAX_SCALE_FACTOR), MIN_SCALE_FACTOR)
+                if (scrollView.height > defaultRowHeight * newFactor * 24) {
+                    newFactor = scrollView.height / 24f / defaultRowHeight
+                }
+
+                if (Math.abs(newFactor - prevScaleFactor) > MIN_SCALE_DIFFERENCE) {
+                    prevScaleFactor = newFactor
+                    config.weeklyViewItemHeightMultiplier = newFactor
+                    updateViewScale()
+                    listener?.updateRowHeight(rowHeight.toInt())
+
+                    val targetY = rowHeightsAtScale * rowHeight - scaleCenterPercent * getVisibleHeight()
+                    scrollView.scrollTo(0, targetY.toInt())
+                }
+                return super.onScale(detector)
+            }
+
+            override fun onScaleBegin(detector: ScaleGestureDetector): Boolean {
+                scaleCenterPercent = detector.focusY / scrollView.height
+                rowHeightsAtScale = (scrollView.scrollY + scaleCenterPercent * getVisibleHeight()) / rowHeight
+                scrollView.isScrollable = false
+                prevScaleSpanY = detector.currentSpanY
+                prevScaleFactor = config.weeklyViewItemHeightMultiplier
+                wasScaled = true
+                screenHeight = context!!.realScreenSize.y
+                return super.onScaleBegin(detector)
+            }
+        })
+    }
+
+    private fun getVisibleHeight(): Float {
+        val fullContentHeight = rowHeight * 24
+        val visibleRatio = scrollView.height / fullContentHeight
+        return fullContentHeight * visibleRatio
     }
 
     override fun updateWeeklyCalendar(events: ArrayList<Event>) {
@@ -234,16 +289,32 @@ class WeekFragment : Fragment(), WeeklyCalendar {
         }
 
         lastHash = newHash
-        this.events = events
 
         activity!!.runOnUiThread {
             if (context != null && activity != null && isAdded) {
-                addEvents()
+                val replaceDescription = config.replaceDescription
+                val sorted = events.sortedWith(
+                        compareBy<Event> { it.startTS }.thenBy { it.endTS }.thenBy { it.title }.thenBy { if (replaceDescription) it.location else it.description }
+                ).toMutableList() as ArrayList<Event>
+
+                currEvents = sorted
+                addEvents(sorted)
             }
         }
     }
 
-    private fun addEvents() {
+    private fun updateViewScale() {
+        rowHeight = context?.getWeeklyViewItemHeight() ?: return
+
+        val oneDp = context!!.resources.getDimension(R.dimen.one_dp).toInt()
+        val fullHeight = Math.max(rowHeight.toInt() * 24, scrollView.height + oneDp)
+        scrollView.layoutParams.height = fullHeight - oneDp
+        mView.week_horizontal_grid_holder.layoutParams.height = fullHeight
+        mView.week_events_columns_holder.layoutParams.height = fullHeight
+        addEvents(currEvents)
+    }
+
+    private fun addEvents(events: ArrayList<Event>) {
         initGrid()
         allDayHolders.clear()
         allDayRows.clear()
@@ -253,15 +324,13 @@ class WeekFragment : Fragment(), WeeklyCalendar {
 
         addNewLine()
 
-        val fullHeight = mRes.getDimension(R.dimen.weekly_view_events_height)
-        val minuteHeight = fullHeight / (24 * 60)
-        val minimalHeight = mRes.getDimension(R.dimen.weekly_view_minimal_event_height).toInt()
-        val density = Math.round(context!!.resources.displayMetrics.density)
+        val minuteHeight = rowHeight / 60
+        val minimalHeight = res.getDimension(R.dimen.weekly_view_minimal_event_height).toInt()
+        val density = Math.round(res.displayMetrics.density)
 
         var hadAllDayEvent = false
-        val replaceDescription = mConfig.replaceDescription
-        val sorted = events.sortedWith(compareBy<Event> { it.startTS }.thenBy { it.endTS }.thenBy { it.title }.thenBy { if (replaceDescription) it.location else it.description })
-        for (event in sorted) {
+
+        for (event in events) {
             val startDateTime = Formatter.getDateTimeFromTS(event.startTS)
             val endDateTime = Formatter.getDateTimeFromTS(event.endTS)
             if (!event.getIsAllDay() && Formatter.getDayCodeFromDateTime(startDateTime) == Formatter.getDayCodeFromDateTime(endDateTime)) {
@@ -279,14 +348,14 @@ class WeekFragment : Fragment(), WeeklyCalendar {
             }
         }
 
-        for (event in sorted) {
+        for (event in events) {
             val startDateTime = Formatter.getDateTimeFromTS(event.startTS)
             val endDateTime = Formatter.getDateTimeFromTS(event.endTS)
             if (event.getIsAllDay() || Formatter.getDayCodeFromDateTime(startDateTime) != Formatter.getDayCodeFromDateTime(endDateTime)) {
                 hadAllDayEvent = true
                 addAllDayEvent(event)
             } else {
-                val dayOfWeek = startDateTime.plusDays(if (mConfig.isSundayFirst) 1 else 0).dayOfWeek - 1
+                val dayOfWeek = startDateTime.plusDays(if (config.isSundayFirst) 1 else 0).dayOfWeek - 1
                 val layout = getColumnWithId(dayOfWeek)
 
                 val startMinutes = startDateTime.minuteOfDay
@@ -343,7 +412,11 @@ class WeekFragment : Fragment(), WeeklyCalendar {
                             }
                         }
 
-                        minHeight = if (event.startTS == event.endTS) minimalHeight else (duration * minuteHeight).toInt() - 1
+                        minHeight = if (event.startTS == event.endTS) {
+                            minimalHeight
+                        } else {
+                            (duration * minuteHeight).toInt() - 1
+                        }
                     }
                     setOnClickListener {
                         Intent(context, EventActivity::class.java).apply {
@@ -377,12 +450,11 @@ class WeekFragment : Fragment(), WeeklyCalendar {
                 mView.week_events_holder.removeView(currentTimeView)
             }
 
-            currentTimeView = (inflater.inflate(R.layout.week_now_marker, null, false) as ImageView)
-            currentTimeView!!.apply {
+            currentTimeView = (inflater.inflate(R.layout.week_now_marker, null, false) as ImageView).apply {
                 applyColorFilter(primaryColor)
                 mView.week_events_holder.addView(this, 0)
                 val extraWidth = (todayColumn.width * 0.3).toInt()
-                val markerHeight = mRes.getDimension(R.dimen.weekly_view_now_height).toInt()
+                val markerHeight = res.getDimension(R.dimen.weekly_view_now_height).toInt()
                 (layoutParams as RelativeLayout.LayoutParams).apply {
                     width = todayColumn.width + extraWidth
                     height = markerHeight
@@ -396,7 +468,7 @@ class WeekFragment : Fragment(), WeeklyCalendar {
     private fun checkTopHolderHeight() {
         mView.week_top_holder.onGlobalLayout {
             if (isFragmentVisible && activity != null && !mWasDestroyed) {
-                mListener?.updateHoursTopMargin(mView.week_top_holder.height)
+                listener?.updateHoursTopMargin(mView.week_top_holder.height)
             }
         }
     }
@@ -418,11 +490,11 @@ class WeekFragment : Fragment(), WeeklyCalendar {
             val startDateTime = Formatter.getDateTimeFromTS(event.startTS)
             val endDateTime = Formatter.getDateTimeFromTS(event.endTS)
 
-            val minTS = Math.max(startDateTime.seconds(), mWeekTimestamp)
-            val maxTS = Math.min(endDateTime.seconds(), mWeekTimestamp + WEEK_SECONDS)
+            val minTS = Math.max(startDateTime.seconds(), weekTimestamp)
+            val maxTS = Math.min(endDateTime.seconds(), weekTimestamp + WEEK_SECONDS)
 
             // fix a visual glitch with all-day events or events lasting multiple days starting at midnight on monday, being shown the previous week too
-            if (minTS == maxTS && (minTS - mWeekTimestamp == WEEK_SECONDS.toLong())) {
+            if (minTS == maxTS && (minTS - weekTimestamp == WEEK_SECONDS.toLong())) {
                 return
             }
 
@@ -430,7 +502,7 @@ class WeekFragment : Fragment(), WeeklyCalendar {
             val numDays = Days.daysBetween(Formatter.getDateTimeFromTS(minTS).toLocalDate(), Formatter.getDateTimeFromTS(maxTS).toLocalDate()).days
             val daysCnt = if (numDays == 1 && isStartTimeDay) 0 else numDays
             val startDateTimeInWeek = Formatter.getDateTimeFromTS(minTS)
-            val firstDayIndex = (startDateTimeInWeek.dayOfWeek - if (mConfig.isSundayFirst) 0 else 1) % 7
+            val firstDayIndex = (startDateTimeInWeek.dayOfWeek - if (config.isSundayFirst) 0 else 1) % 7
 
             var doesEventFit: Boolean
             val cnt = allDayRows.size - 1
@@ -488,22 +560,27 @@ class WeekFragment : Fragment(), WeeklyCalendar {
         mView.week_top_holder.onGlobalLayout {
             if (activity != null && !mWasDestroyed) {
                 if (isFragmentVisible) {
-                    mListener?.updateHoursTopMargin(mView.week_top_holder.height)
+                    listener?.updateHoursTopMargin(mView.week_top_holder.height)
                 }
 
                 if (!wasExtraHeightAdded) {
-                    maxScrollY += mView.week_all_day_holder.height
                     wasExtraHeightAdded = true
                 }
             }
         }
     }
 
-    private fun getColumnWithId(id: Int) = mView.findViewById<ViewGroup>(mRes.getIdentifier("week_column_$id", "id", context!!.packageName))
+    private fun getColumnWithId(id: Int) = mView.findViewById<ViewGroup>(res.getIdentifier("week_column_$id", "id", context!!.packageName))
 
     fun updateScrollY(y: Int) {
         if (wasFragmentInit) {
-            mScrollView.scrollY = y
+            scrollView.scrollY = y
+        }
+    }
+
+    fun updateNotVisibleViewScaleLevel() {
+        if (!isFragmentVisible) {
+            updateViewScale()
         }
     }
 }